:sparkles:大幅度修改参数数据结构

+++ src/main/java/com/dong/djudge/entity/judge/CodeSetting.java
@@ -11,15 +11,15 @@ public class CodeSetting {
     /**
      * 最大运行时间
      */
-    Long maxTime;
+    Integer maxTime;
     /**
      * 最大运行内存
      */
-    Long maxMemory;
+    Integer maxMemory;
     /**
      * 最大栈大小
      */
-    Long maxStack;
+    Integer maxStack;
     /**
      * 最大输出大小
      */
